requirements.txt
# av
accelerate
config
# cloudpickle
diffusers
einops
# fvcore
# imageio
# matplotlib
numpy
omegaconf
onnxruntime
opencv-python
pandas
pillow
# pycocotools
pyyaml
# psutil
peft
regex==2024.5.15
safetensors
scikit-image
scipy
timm
tokenizers
torch
torchaudio
# torchmetrics
torchvision
tqdm
transformers
fastapi
uvicorn
huggingface-hub
# triton